pyproject.toml
[project]
name = "fast-doctolib"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.116.2",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.25",
]
